watch DOM for changes to reload scarcity counters
this adds support for elements created asynchronously e.g. with React, Vue, Angular, etc.

When ever the DOM of the body changes all counters are updated. This can lead to regular updates of the counters. Use the timeout attribute to set a limit for these reloads. e.g. 5000 to update the counters only every 5 sections regardless of the numbers of DOM changes.